requirements.txt
pyrogram
tgcrypto
python-dotenv 
sqlalchemy==1.3.20
psycopg2-binary
